現在ユーザーのaccount情報取得

+++ front/src/session/index.ts
@@ -1,16 +1,24 @@
 export const setSession = (key: string, value: string) => {
-  if(!value) return;
-  localStorage.setItem(key, value);
+  if (typeof window !== "undefined" && value) {
+    localStorage.setItem(key, value);
+  }
 };
 
 export const getSession = (key: string) => {
-  return localStorage.getItem(key);
+  if (typeof window !== "undefined") {
+    return localStorage.getItem(key);
+  }
+  return null;
 };
 
 export const removeSession = (key: string) => {
-  localStorage.removeItem(key);
+  if (typeof window !== "undefined") {
+    localStorage.removeItem(key);
+  }
 };
 
 export const clearSession = () => {
-  localStorage.clear();
+  if (typeof window !== "undefined") {
+    localStorage.clear();
+  }
 };
